Refactoring macro tests to make them more modular (so I can I add more tests)
